rename sdk folders
