+++ sdk/python/utils.py
