+++ tests/go.mod
@@ -9,7 +9,7 @@ replace (
 	github.com/ease-lab/vhive-xdt/proto/upXDT => ../proto/upXDT
 	github.com/ease-lab/vhive-xdt/queue-proxy/dQP => ../queue-proxy/dQP
 	github.com/ease-lab/vhive-xdt/queue-proxy/sQP => ../queue-proxy/sQP
-	github.com/ease-lab/vhive-xdt/sdk/go_sdk => ../sdk/go_sdk
+	github.com/ease-lab/vhive-xdt/sdk/golang => ../sdk/golang
 	github.com/ease-lab/vhive-xdt/transport => ../transport
 	github.com/ease-lab/vhive-xdt/utils => ../utils
 )
@@ -19,7 +19,7 @@ require (
 	github.com/containerd/containerd v1.5.2
 	github.com/ease-lab/vhive-xdt/queue-proxy/dQP v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/vhive-xdt/queue-proxy/sQP v0.0.0-00010101000000-000000000000
-	github.com/ease-lab/vhive-xdt/sdk/go_sdk v0.0.0-00010101000000-000000000000
+	github.com/ease-lab/vhive-xdt/sdk/golang v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/vhive-xdt/utils v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/vhive/utils/tracing/go v0.0.0-20210624210547-e0cd5d053491
 	github.com/sirupsen/logrus v1.8.1
