+++ user-functions/fx/go.mod
@@ -5,13 +5,13 @@ go 1.16
 replace (
 	github.com/ease-lab/vhive-xdt/proto/downXDT => ../../proto/downXDT
 	github.com/ease-lab/vhive-xdt/proto/upXDT => ../../proto/upXDT
-	github.com/ease-lab/vhive-xdt/sdk/go_sdk => ../../sdk/go_sdk
+	github.com/ease-lab/vhive-xdt/sdk/golang => ../../sdk/golang
 	github.com/ease-lab/vhive-xdt/utils => ../../utils
 )
 
 require (
 	github.com/containerd/containerd v1.5.2
-	github.com/ease-lab/vhive-xdt/sdk/go_sdk v0.0.0-00010101000000-000000000000
+	github.com/ease-lab/vhive-xdt/sdk/golang v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/vhive-xdt/utils v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/vhive/examples/protobuf/helloworld v0.0.0-20210624210547-e0cd5d053491
 	github.com/sirupsen/logrus v1.8.1
